# parking-managment-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Parking Management API
  description: |
    API REST para el sistema de gestión de estacionamientos.
    
    Esta API permitirá:
    - Gestión de usuarios y autenticación
    - Administración de estacionamientos y espacios
    - Reservas programadas y walk-in
    - Sistema de reseñas e incidentes
    - Gestión de vehículos y precios
    
    **Autenticación:**
    Utiliza JWT (JSON Web Tokens) para la autenticación. 
    Obtén un token mediante el endpoint /auth/login y úsalo en el header Authorization.
    
    **Formato del token:**
    `Authorization: Bearer <tu-jwt-token>`
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@parkingmanagement.com
    url: https://github.com//parking-management-api
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8081/api
    description: Servidor de desarrollo

security:
  - bearerAuth: []

# Esquemas de seguridad
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authentication. Formato Bearer <token>
  
  # Esquemas de datos
  schemas:
    User:
      type: object
      description: Modelo de usuario del sistema
      required:
        - firstName
        - lastName
        - email
        - passwordHash
      properties:
        id:
          type: integer
          format: int64
          description: ID único del usuario
          example: 1
          readOnly: true
        firstName:
          type: string
          maxLength: 100
          description: Nombre del usuario
          example: Juan
        lastName:
          type: string
          maxLength: 100
          description: Apellido del usuario
          example: Pérez
        email:
          type: string
          format: email
          description: Email del usuario (único)
          example: juan.perez@gmail.com
        passwordHash:
          type: string
          minLength: 6
          description: Contraseña del usuario
          example: password123
          writeOnly: true
        imageUrl:
          type: string
          format: uri
          description: URL de la imagen de perfil
          example: https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=150
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación del usuario
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Fecha de última actualización
          readOnly: true

    UserCreate:
      type: object
      description: Datos para crear un nuevo usuario
      required:
        - firstName
        - lastName
        - email
        - passwordHash
      properties:
        firstName:
          type: string
          maxLength: 100
          description: Nombre del usuario
          example: Juan
        lastName:
          type: string
          maxLength: 100
          description: Apellido del usuario
          example: Pérez
        email:
          type: string
          format: email
          description: Email del usuario (único)
          example: juan.perez@gmail.com
        passwordHash:
          type: string
          minLength: 6
          description: Contraseña del usuario
          example: password123
        imageUrl:
          type: string
          format: uri
          description: URL de la imagen de perfil
          example: https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=150

    UserUpdate:
      type: object
      description: Datos para actualizar un usuario
      properties:
        firstName:
          type: string
          maxLength: 100
          description: Nombre del usuario
          example: Juan
        lastName:
          type: string
          maxLength: 100
          description: Apellido del usuario
          example: Pérez
        email:
          type: string
          format: email
          description: Email del usuario (único)
          example: juan.perez@gmail.com
        passwordHash:
          type: string
          minLength: 6
          description: Contraseña del usuario
          example: password123
        imageUrl:
          type: string
          format: uri
          description: URL de la imagen de perfil
          example: https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=150

    ErrorResponse:
      type: object
      description: Respuesta de error estándar
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp del error
        status:
          type: integer
          description: Código de estado HTTP
        error:
          type: string
          description: Descripción del error HTTP
        message:
          type: string
          description: Mensaje detallado del error
        path:
          type: string
          description: Ruta donde ocurrió el error

# Rutas de la API
paths:
  /users:
    get:
      tags:
        - Usuarios
      summary: Obtener todos los usuarios
      description: Retorna una lista con todos los usuarios registrados en el sistema
      operationId: getAllUsers
      responses:
        '200':
          description: Lista de usuarios obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              examples:
                usuarios_ejemplo:
                  summary: Ejemplo de lista de usuarios
                  value:
                    - id: 1
                      firstName: Juan
                      lastName: Pérez
                      email: juan.perez@gmail.com
                      imageUrl: https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=150
                      createdAt: "2024-12-19T10:00:00Z"
                      updatedAt: "2024-12-19T10:00:00Z"
                    - id: 2
                      firstName: María
                      lastName: González
                      email: maria.gonzalez@outlook.com
                      imageUrl: https://images.unsplash.com/photo-1494790108755-2616b612b786?w=150
                      createdAt: "2024-12-19T10:05:00Z"
                      updatedAt: "2024-12-19T10:05:00Z"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Usuarios
      summary: Crear nuevo usuario
      description: Crea un nuevo usuario en el sistema con toda la información requerida
      operationId: createUser
      requestBody:
        description: Información del usuario a crear
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples:
              usuario_nuevo:
                summary: Ejemplo de usuario nuevo
                value:
                  firstName: Carlos
                  lastName: Rodríguez
                  email: carlos.rodriguez@hotmail.com
                  passwordHash: password123
                  imageUrl: https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=150
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                usuario_creado:
                  summary: Usuario creado exitosamente
                  value:
                    id: 3
                    firstName: Carlos
                    lastName: Rodríguez
                    email: carlos.rodriguez@hotmail.com
                    imageUrl: https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=150
                    createdAt: "2024-12-19T10:10:00Z"
                    updatedAt: "2024-12-19T10:10:00Z"
        '400':
          description: Datos inválidos o email ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                email_duplicado:
                  summary: Email ya existe
                  value:
                    timestamp: "2024-12-19T10:10:00Z"
                    status: 400
                    error: Bad Request
                    message: El email ya está registrado
                    path: /api/users

  /users/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener usuario por ID
      description: Retorna un usuario específico basado en su ID único
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: ID único del usuario
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Usuario encontrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                usuario_encontrado:
                  summary: Usuario encontrado
                  value:
                    id: 1
                    firstName: Juan
                    lastName: Pérez
                    email: juan.perez@gmail.com
                    imageUrl: https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=150
                    createdAt: "2024-12-19T10:00:00Z"
                    updatedAt: "2024-12-19T10:00:00Z"
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                usuario_no_encontrado:
                  summary: Usuario no encontrado
                  value:
                    timestamp: "2024-12-19T10:15:00Z"
                    status: 404
                    error: Not Found
                    message: Usuario con ID 999 no encontrado
                    path: /api/users/999

    put:
      tags:
        - Usuarios
      summary: Actualizar usuario
      description: Actualiza los datos de un usuario existente
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: ID único del usuario
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        description: Información del usuario a actualizar
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
            examples:
              usuario_actualizar:
                summary: Datos para actualizar
                value:
                  firstName: Juan Carlos
                  lastName: Pérez García
                  email: juancarlos.perez@gmail.com
                  imageUrl: https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=200
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Usuarios
      summary: Eliminar usuario
      description: Elimina un usuario del sistema
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID único del usuario
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Usuario eliminado exitosamente
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/search:
    get:
      tags:
        - Usuarios
      summary: Buscar usuarios
      description: Busca usuarios por nombre, apellido o email usando un término de búsqueda
      operationId: searchUsers
      parameters:
        - name: q
          in: query
          description: Término de búsqueda (nombre, apellido o email)
          required: true
          schema:
            type: string
            example: juan
      responses:
        '200':
          description: Búsqueda realizada exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              examples:
                busqueda_juan:
                  summary: Resultados para "juan"
                  value:
                    - id: 1
                      firstName: Juan
                      lastName: Pérez
                      email: juan.perez@gmail.com
                      imageUrl: https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=150
                      createdAt: "2024-12-19T10:00:00Z"
                      updatedAt: "2024-12-19T10:00:00Z"

# Tags para agrupar endpoints
tags:
  - name: Usuarios
    description: Operaciones relacionadas con la gestión de usuarios 